Calculate readyClusters from bundledeployments directly
This should work with overrideTargets
