+++ integrationtests/gitjob/controller/controller_test.go
@@ -2,16 +2,12 @@ package controller
 
 import (
 	"context"
-	"encoding/hex"
 	"fmt"
-	"math/rand"
 	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/onsi/gomega/gcustom"
-	gomegatypes "github.com/onsi/gomega/types"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
@@ -23,7 +19,6 @@ import (
 
 	"github.com/rancher/wrangler/v3/pkg/genericcondition"
 
-	"github.com/rancher/fleet/integrationtests/utils"
 	"github.com/rancher/fleet/internal/names"
 	v1alpha1 "github.com/rancher/fleet/pkg/apis/fleet.cattle.io/v1alpha1"
 )
@@ -794,255 +789,3 @@ var _ = Describe("GitJob controller", func() {
 		})
 	})
 })
-
-var _ = Describe("GitRepo", func() {
-	var (
-		gitrepo     *v1alpha1.GitRepo
-		gitrepoName string
-	)
-
-	BeforeEach(func() {
-		var err error
-		namespace, err = utils.NewNamespaceName()
-		Expect(err).ToNot(HaveOccurred())
-		ns := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
-		Expect(k8sClient.Create(ctx, ns)).ToNot(HaveOccurred())
-
-		p := make([]byte, 12)
-		s := rand.New(rand.NewSource(GinkgoRandomSeed())) // nolint:gosec // non-crypto usage
-		if _, err := s.Read(p); err != nil {
-			panic(err)
-		}
-		gitrepoName = fmt.Sprintf("test-gitrepo-%.12s", hex.EncodeToString(p))
-
-		gitrepo = &v1alpha1.GitRepo{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      gitrepoName,
-				Namespace: namespace,
-			},
-			Spec: v1alpha1.GitRepoSpec{
-				Repo: "https://github.com/rancher/fleet-test-data/not-found",
-			},
-		}
-
-		DeferCleanup(func() {
-			Expect(k8sClient.Delete(ctx, gitrepo)).ToNot(HaveOccurred())
-		})
-	})
-
-	When("creating a gitrepo", func() {
-		JustBeforeEach(func() {
-			expectedCommit = commit
-			err := k8sClient.Create(ctx, gitrepo)
-			Expect(err).NotTo(HaveOccurred())
-		})
-
-		It("updates the gitrepo status", func() {
-			org := gitrepo.ResourceVersion
-			Eventually(func(g Gomega) {
-				_ = k8sClient.Get(ctx, types.NamespacedName{Name: gitrepoName, Namespace: namespace}, gitrepo)
-				g.Expect(gitrepo.ResourceVersion > org).To(BeTrue())
-				g.Expect(gitrepo.Status.Display.ReadyBundleDeployments).To(Equal("0/0"))
-				g.Expect(gitrepo.Status.Display.Error).To(BeFalse())
-				g.Expect(len(gitrepo.Status.Conditions)).To(Equal(5))
-				g.Expect(checkCondition(gitrepo, "GitPolling", corev1.ConditionTrue, "")).To(BeTrue())
-				g.Expect(checkCondition(gitrepo, "Reconciling", corev1.ConditionTrue, "")).To(BeTrue())
-				g.Expect(checkCondition(gitrepo, "Stalled", corev1.ConditionFalse, "")).To(BeTrue())
-				g.Expect(checkCondition(gitrepo, "Ready", corev1.ConditionTrue, "")).To(BeTrue())
-				g.Expect(checkCondition(gitrepo, "Accepted", corev1.ConditionTrue, "")).To(BeTrue())
-			}).Should(Succeed())
-		})
-	})
-})
-
-var _ = Describe("GitRepo Status Fields", func() {
-	var (
-		gitrepo *v1alpha1.GitRepo
-		bd      *v1alpha1.BundleDeployment
-	)
-
-	BeforeEach(func() {
-		var err error
-		namespace, err = utils.NewNamespaceName()
-		Expect(err).ToNot(HaveOccurred())
-
-		ns := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
-		Expect(k8sClient.Create(ctx, ns)).ToNot(HaveOccurred())
-
-		DeferCleanup(func() {
-			Expect(k8sClient.Delete(ctx, ns)).ToNot(HaveOccurred())
-		})
-	})
-
-	When("Bundle changes", func() {
-		BeforeEach(func() {
-			cluster, err := utils.CreateCluster(ctx, k8sClient, "cluster", namespace, nil, namespace)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(cluster).To(Not(BeNil()))
-			targets := []v1alpha1.BundleTarget{
-				{
-					BundleDeploymentOptions: v1alpha1.BundleDeploymentOptions{
-						TargetNamespace: "targetNs",
-					},
-					Name:        "cluster",
-					ClusterName: "cluster",
-				},
-			}
-			bundle, err := utils.CreateBundle(ctx, k8sClient, "name", namespace, targets, targets)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(bundle).To(Not(BeNil()))
-
-			gitrepo = &v1alpha1.GitRepo{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-gitrepo",
-					Namespace: namespace,
-				},
-				Spec: v1alpha1.GitRepoSpec{
-					Repo: "https://github.com/rancher/fleet-test-data/not-found",
-				},
-			}
-			err = k8sClient.Create(ctx, gitrepo)
-			Expect(err).NotTo(HaveOccurred())
-
-			bd = &v1alpha1.BundleDeployment{}
-			Eventually(func() bool {
-				err = k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: "name"}, bd)
-				return err == nil
-			}).Should(BeTrue())
-		})
-
-		It("updates the status fields", func() {
-			bundle := &v1alpha1.Bundle{}
-			By("Receiving a bundle update")
-			Eventually(func() error {
-				err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: "name"}, bundle)
-				Expect(err).ToNot(HaveOccurred())
-				bundle.Labels["fleet.cattle.io/repo-name"] = gitrepo.Name
-				return k8sClient.Update(ctx, bundle)
-			}).ShouldNot(HaveOccurred())
-			Expect(bundle.Status.Summary.Ready).ToNot(Equal(1))
-
-			By("Updating the GitRepo status to not ready")
-			err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: gitrepo.Name}, gitrepo)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(gitrepo.Status.Summary.Ready).To(Equal(0))
-
-			// This simulates what the bundle deployment reconciler would do.
-			By("Updating the Bundle deployment status to ready")
-			bd := &v1alpha1.BundleDeployment{}
-			Eventually(func() error {
-				err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: "name"}, bd)
-				if err != nil {
-					return err
-				}
-				bd.Status.Display.State = "Ready"
-				bd.Status.AppliedDeploymentID = bd.Spec.DeploymentID
-				bd.Status.Ready = true
-				bd.Status.NonModified = true
-				return k8sClient.Status().Update(ctx, bd)
-			}).ShouldNot(HaveOccurred())
-
-			By("Updating the GitRepo status to ready")
-			Eventually(func() bool {
-				err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: "name"}, bundle)
-				Expect(err).NotTo(HaveOccurred())
-				return bundle.Status.Summary.Ready == 1
-			}).Should(BeTrue())
-			err = k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: gitrepo.Name}, gitrepo)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(gitrepo.Status.Summary.Ready).To(Equal(1))
-
-			By("Deleting a bundle")
-			err = k8sClient.Delete(ctx, bundle)
-			Expect(err).ToNot(HaveOccurred())
-
-			Eventually(func(g Gomega) {
-				err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: gitrepo.Name}, gitrepo)
-				g.Expect(err).ToNot(HaveOccurred())
-				g.Expect(gitrepo.Status.Summary.Ready).To(Equal(0))
-				g.Expect(gitrepo.Status.Summary.DesiredReady).To(Equal(0))
-				g.Expect(gitrepo.Status.Display.ReadyBundleDeployments).To(Equal("0/0"))
-			}).Should(Succeed())
-		})
-	})
-})
-
-func simulateIncreaseForceSyncGeneration(gitRepo v1alpha1.GitRepo) error {
-	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		var gitRepoFromCluster v1alpha1.GitRepo
-		err := k8sClient.Get(ctx, types.NamespacedName{Name: gitRepo.Name, Namespace: gitRepo.Namespace}, &gitRepoFromCluster)
-		if err != nil {
-			return err
-		}
-		gitRepoFromCluster.Spec.ForceSyncGeneration++
-		return k8sClient.Update(ctx, &gitRepoFromCluster)
-	})
-}
-
-func setGitRepoWebhookCommit(gitRepo v1alpha1.GitRepo, commit string) error {
-	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		var gitRepoFromCluster v1alpha1.GitRepo
-		err := k8sClient.Get(ctx, types.NamespacedName{Name: gitRepo.Name, Namespace: gitRepo.Namespace}, &gitRepoFromCluster)
-		if err != nil {
-			return err
-		}
-		gitRepoFromCluster.Status.WebhookCommit = commit
-		return k8sClient.Status().Update(ctx, &gitRepoFromCluster)
-	})
-}
-
-func createGitRepo(gitRepoName string) v1alpha1.GitRepo {
-	return v1alpha1.GitRepo{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      gitRepoName,
-			Namespace: gitRepoNamespace,
-		},
-		Spec: v1alpha1.GitRepoSpec{
-			Repo: repo,
-		},
-	}
-}
-
-func createGitRepoWithDisablePolling(gitRepoName string) v1alpha1.GitRepo {
-	return v1alpha1.GitRepo{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      gitRepoName,
-			Namespace: gitRepoNamespace,
-		},
-		Spec: v1alpha1.GitRepoSpec{
-			Repo:           repo,
-			DisablePolling: true,
-			Branch:         stableCommitBranch,
-		},
-	}
-}
-
-func waitDeleteGitrepo(gitRepo v1alpha1.GitRepo) {
-	err := k8sClient.Delete(ctx, &gitRepo)
-	Expect(err).ToNot(HaveOccurred())
-	Eventually(func() bool {
-		var gitRepoFromCluster v1alpha1.GitRepo
-		err := k8sClient.Get(ctx, types.NamespacedName{Name: gitRepo.Name, Namespace: gitRepo.Namespace}, &gitRepoFromCluster)
-		return errors.IsNotFound(err)
-	}).Should(BeTrue())
-}
-
-func beOwnedBy(expected interface{}) gomegatypes.GomegaMatcher {
-	return gcustom.MakeMatcher(func(meta metav1.ObjectMeta) (bool, error) {
-		ref, ok := expected.(metav1.OwnerReference)
-		if !ok {
-			return false, fmt.Errorf("beOwnedBy matcher expects metav1.OwnerReference")
-		}
-
-		for _, or := range meta.OwnerReferences {
-			if or.Kind == ref.Kind && or.APIVersion == ref.APIVersion && or.Name == ref.Name {
-				return true, nil
-			}
-		}
-
-		return false, nil
-	}).WithTemplate(
-		"Expected:\n{{.FormattedActual}}\n{{.To}} contain owner reference " +
-			"matching Kind, APIVersion and Name of \n{{format .Data 1}}",
-	).WithTemplateData(expected)
-}
