+++ integrationtests/gitjob/controller/suite_test.go
@@ -3,12 +3,15 @@ package controller
 import (
 	"bytes"
 	"context"
+	"fmt"
 	"path/filepath"
 	"testing"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/gcustom"
+	gomegatypes "github.com/onsi/gomega/types"
 	"github.com/reugn/go-quartz/quartz"
 	"go.uber.org/mock/gomock"
 
@@ -20,9 +23,13 @@ import (
 	v1alpha1 "github.com/rancher/fleet/pkg/apis/fleet.cattle.io/v1alpha1"
 	"github.com/rancher/fleet/pkg/git/mocks"
 
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
+	"k8s.io/client-go/util/retry"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
@@ -137,3 +144,83 @@ var _ = AfterSuite(func() {
 	cancel()
 	Expect(testEnv.Stop()).ToNot(HaveOccurred())
 })
+
+func simulateIncreaseForceSyncGeneration(gitRepo v1alpha1.GitRepo) error {
+	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		var gitRepoFromCluster v1alpha1.GitRepo
+		err := k8sClient.Get(ctx, types.NamespacedName{Name: gitRepo.Name, Namespace: gitRepo.Namespace}, &gitRepoFromCluster)
+		if err != nil {
+			return err
+		}
+		gitRepoFromCluster.Spec.ForceSyncGeneration++
+		return k8sClient.Update(ctx, &gitRepoFromCluster)
+	})
+}
+
+func setGitRepoWebhookCommit(gitRepo v1alpha1.GitRepo, commit string) error {
+	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		var gitRepoFromCluster v1alpha1.GitRepo
+		err := k8sClient.Get(ctx, types.NamespacedName{Name: gitRepo.Name, Namespace: gitRepo.Namespace}, &gitRepoFromCluster)
+		if err != nil {
+			return err
+		}
+		gitRepoFromCluster.Status.WebhookCommit = commit
+		return k8sClient.Status().Update(ctx, &gitRepoFromCluster)
+	})
+}
+
+func createGitRepo(gitRepoName string) v1alpha1.GitRepo {
+	return v1alpha1.GitRepo{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      gitRepoName,
+			Namespace: gitRepoNamespace,
+		},
+		Spec: v1alpha1.GitRepoSpec{
+			Repo: repo,
+		},
+	}
+}
+
+func createGitRepoWithDisablePolling(gitRepoName string) v1alpha1.GitRepo {
+	return v1alpha1.GitRepo{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      gitRepoName,
+			Namespace: gitRepoNamespace,
+		},
+		Spec: v1alpha1.GitRepoSpec{
+			Repo:           repo,
+			DisablePolling: true,
+			Branch:         stableCommitBranch,
+		},
+	}
+}
+
+func waitDeleteGitrepo(gitRepo v1alpha1.GitRepo) {
+	err := k8sClient.Delete(ctx, &gitRepo)
+	Expect(err).ToNot(HaveOccurred())
+	Eventually(func() bool {
+		var gitRepoFromCluster v1alpha1.GitRepo
+		err := k8sClient.Get(ctx, types.NamespacedName{Name: gitRepo.Name, Namespace: gitRepo.Namespace}, &gitRepoFromCluster)
+		return errors.IsNotFound(err)
+	}).Should(BeTrue())
+}
+
+func beOwnedBy(expected interface{}) gomegatypes.GomegaMatcher {
+	return gcustom.MakeMatcher(func(meta metav1.ObjectMeta) (bool, error) {
+		ref, ok := expected.(metav1.OwnerReference)
+		if !ok {
+			return false, fmt.Errorf("beOwnedBy matcher expects metav1.OwnerReference")
+		}
+
+		for _, or := range meta.OwnerReferences {
+			if or.Kind == ref.Kind && or.APIVersion == ref.APIVersion && or.Name == ref.Name {
+				return true, nil
+			}
+		}
+
+		return false, nil
+	}).WithTemplate(
+		"Expected:\n{{.FormattedActual}}\n{{.To}} contain owner reference " +
+			"matching Kind, APIVersion and Name of \n{{format .Data 1}}",
+	).WithTemplateData(expected)
+}
